Income Distribution by Community Area
This chart highlights in red those community areas that belong to the 42nd Ward, which is where most of the restaurant licenses are concentrated. Thus, this sole graph shows that there is a direct relationship between the concentration of licenses and the concentration of wealth by per capita income, which suggests that wealthy areas are better for restaurants. 

This graph contributes to answering question 3.